dom/workers/test/serviceworkers/test_origin_after_redirect_cached.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test the origin of a redirected response from a service worker</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
<iframe></iframe>
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var iframe;
  function runTest() {
    iframe = document.querySelector("iframe");
    iframe.src = "/tests/dom/workers/test/serviceworkers/fetch/origin/register.html";
    var win;
    window.onmessage = function(e) {
      if (e.data.status == "ok") {
        ok(e.data.result, e.data.message);
      } else if (e.data.status == "registrationdone") {
        win = window.open("/tests/dom/workers/test/serviceworkers/fetch/origin/index-cached.sjs", "mywindow", "width=100,height=100");
      } else if (e.data.status == "domain") {
        is(e.data.data, "example.org", "Correct domain expected");
      } else if (e.data.status == "origin") {
        is(e.data.data, "http://example.org", "Correct origin expected");
      } else if (e.data.status == "done") {
        win.close();
        iframe.src = "/tests/dom/workers/test/serviceworkers/fetch/origin/unregister.html";
      } else if (e.data.status == "unregistrationdone") {
        window.onmessage = null;
        ok(true, "Test finished successfully");
        SimpleTest.finish();
      }
    };
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
      ["dom.caches.enabled", true],
    ]}, runTest);
  };
</script>
</pre>
</body>
</html>
